Correct the test case
I had incorrectly assumed that the guardsize was being adjusted
against the page size.  What's actually happening is that
allocate_stack perturbs the stack size by a page if it is a multiple
of 64k to avoid aliasing conflicts that cause a performance drop on P4
and later.  So now I perturb the stacksize request by a page in
advance, so that pthread_create does not modify the requested size.