PowerPC: IBM long double fixes
This patch starts the process of supporting powerpc64 little-endian
long double in glibc.  IBM long double is an array of two ieee
doubles, so making union ibm_extended_long_double reflect this fact is
the correct way to access fields of the doubles IMO, rather than
adding a little-endian version of the current union.  That of course
results in a much larger but mechanical patch.